+++ packages/react-route/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @minko-fe/react-route
 
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @minko-fe/react-hook@0.3.3
+
 ## 0.1.28
 
 ### Patch Changes